+++ scripts/allClusterDynamics_faster.py
@@ -24,13 +24,11 @@
 cluster_path = "../ncov_cluster/cluster_profile/"
 
 # Things that write out to cluster_scripts repo (images mostly), use this path:
-figure_path = "../covariants/overall_trends_figures/"
 tables_path = "../covariants/cluster_tables/"
 overall_tables_file = "../covariants/cluster_tables/all_tables.tsv"
 acknowledgement_folder = "../covariants/acknowledgements/"
 acknowledgement_folder_new = "../covariants/web/public/acknowledgements/"
 web_data_folder = "../covariants/web/data/"
-figure_only_path = "../covariants/figures/"
 # This assumes that `covariants` sites next to `ncov`
 # Otherwise, modify the paths above to put the files wherever you like.
 # (Alternatively just create a folder structure to mirror the above)
@@ -38,7 +36,11 @@ figure_only_path = "../covariants/figures/"
 fmt = "png"  # "pdf"
 grey_color = "#cccccc"  # for "other clusters" of country plots
 
-dated_limit = "2021-03-31" #only works for Q677 currently
+#dated_limit = "2021-03-31" #only works for Q677 currently
+#dated_limit = "2021-06-30"
+#dated_cluster = "21A (Delta)"
+#dated_cluster = "20I (Alpha, V1)"
+dated_cluster = "Q677"
 dated_limit = ""
 
 import pandas as pd
@@ -52,9 +54,7 @@ from matplotlib.patches import Rectangle
 import json
 import matplotlib.patches as mpatches
 from colors_and_countries import *
-from travel_data import *
 from helpers import *
-from paths import *
 from clusters import *
 from bad_sequences import *
 from approx_first_dates import *
@@ -215,6 +215,28 @@ while reask:
         reask = False
 print("These clusters will be run: ", clus_to_run)
 
+# ask user if they want to continue to do the full plotting - mostly we do these days 
+do_country = False
+if "all" in clus_answer:
+    print_answer = input("\nContinue to country plotting? (y/n) (Enter is no): ")
+    if print_answer in ["y", "Y", "yes", "YES", "Yes"]:
+        do_country = True
+else:
+    print("Can't do country plot as aren't doing 'all' clusters")
+
+if do_country == False:
+    print(
+        "You can alway run this step by calling `plot_country_data(clusters, proposed_coun_to_plot, print_files)`"
+    )
+
+do_divisions_country = False
+if "all" in clus_answer:
+    print_answer = input(
+        "\nContinue to USA- & Swiss-specific country plotting? (y/n) (Enter is no): "
+    )
+    if print_answer in ["y", "Y", "yes", "YES", "Yes"]:
+        do_divisions_country = True
+
 start_time = time.time()
 
 ##################################
@@ -289,32 +311,7 @@ json_output = {}
 
 # if running all clusters, clear file so can write again.
 if print_files and "all" in clus_answer:
-    # clean these files so don't append to last run.
-    with open(f"{tables_path}all_tables.md", "w") as fh:
-        fh.write("\n")
-        fh.write(
-            "# Overview of Clusters/Mutations in Europe\n"
-            "[Overview of proportion of clusters in selected countries](country_overview.md)\n\n"
-            "In the graphs below, countries are displayed in the chart if the country has at least 20 sequences present in the cluster.\n\n"
-            "# Mutation Tables and Graphs\n"
-            "- [20E (EU1)](#20e-eu1) _(S:A222V)_ \n"
-            "- [20A.EU2](#20aeu2) _(S:S477N)_ \n"
-            "- [20I/501Y.V1](#20i501yv1) \n"
-            "- [20H/501Y.V2](#20h501yv2) \n"
-            "- [20J/501Y.V3](#20j501yv3) \n"
-            "- [20C/S:452R](#20cs452r) \n"
-            "- [20A/S:439K](#20as439k) \n"
-            "- [20A/S:98F](#20as98f) \n"
-            "- [20C/S:80Y](#20cs80y) \n"
-            "- [20B/S:626S](#20bs626s) \n"
-            "- [20B/S:1122L](#20bs1122l) \n"
-            "- [S:N501](#sn501) \n"
-            "- [S:E484](#se484) \n"
-            "- [S:H69-](#sh69-) \n"
-            "- [S:Q677](#sq677) \n"
-            "- [S:Y453F](#sy453f) \n"
-            "- [S:S477](#ss477) \n\n"
-        )
+    #empty file to write clean
     with open(overall_tables_file, "w") as fh:
         fh.write("\n")
 
@@ -337,7 +334,7 @@ for clus in clus_to_run:
     clus_data = clusters[clus]
 
     if clus == "mink":
-        clus_display = "mink"
+        clus_build_name = "mink"
         clus_data["mink_meta"] = meta[meta["host"].apply(lambda x: x == "Mink")]
         clus_data["wanted_seqs"] = list(clus_data["mink_meta"]["strain"])
 
@@ -347,7 +344,7 @@ for clus in clus_to_run:
         )
 
     else:
-        clus_display = clusters[clus]["build_name"]
+        clus_build_name = clusters[clus]["build_name"]
         snps = clusters[clus]["snps"]
         if "snps2" not in clusters[clus]:
             clus_data["snps2"] = []
@@ -384,6 +381,7 @@ muts["gap_pos"] = muts.nucleotide.fillna('').apply(lambda x: [int(y[1:-1]) for y
 # Figure out what clades are really recognised
 official_clades = list(meta["Nextstrain_clade"].unique())
 
+n_done = 1
 for clus in [x for x in clus_to_run if x != "mink"]:
     clus_data = clusters[clus]
     snps = clus_data["snps"]
@@ -425,6 +423,9 @@ for clus in [x for x in clus_to_run if x != "mink"]:
 
     #dedup
     wanted_seqs = list(set(wanted_seqs))
+    t1a = time.time()
+    print(f"Completed {n_done} out of {len(clus_to_run)}: {round((t1a-t0)/60,1)} min")
+    n_done += 1
 
 
 t1 = time.time()
@@ -435,17 +436,18 @@ print(f"Finding sequences took {round((t1-t0)/60,1)} min to run")
 ##################################
 #### Gather metadata
 t0 = time.time()
+n_done = 1
 for clus in clus_to_run:
-    print(f"\nGathering metadata for cluster {clus}\n")
+    print(f"\nGathering metadata for cluster {clus}, cluster {n_done} out of {len(clus_to_run)}\n")
 
     clus_data = clusters[clus]
     wanted_seqs = clus_data["wanted_seqs"]
-    clus_display = clus_data["build_name"]
+    clus_build_name = clus_data["build_name"]
     display_cluster = clus_data["display_name"]
     clusterlist_output = clus_data["clusterlist_output"]
     out_meta_file = clus_data["out_meta_file"]
 
-    json_output[clus_display] = {}
+    json_output[clus_build_name] = {}
 
     # get metadata for these sequences
     cluster_meta = meta[meta["strain"].isin(wanted_seqs)].copy()
@@ -490,8 +492,10 @@ for clus in clus_to_run:
         cluster_meta.to_csv(out_meta_file, sep="\t", index=False)
 
     # If specified wanted dated Q677, do this
-    if print_files and dated_limit and "Q677" in display_cluster:
-    #if dated_limit and "Q677" in clus_display:
+    #if print_files and dated_limit and "Q677" in display_cluster:
+    # if want a dated limit, specify cluster and limit at top
+    if print_files and dated_limit and dated_cluster in display_cluster:
+    ####### if dated_limit and "Q677" in clus_build_name: # (old)
         build_nam = clusters[clus]["build_name"]
         dated_clus_met = cluster_meta[cluster_meta["date_formatted"].apply(lambda x: x < datetime.datetime.strptime(dated_limit, "%Y-%m-%d"))]
         dated_want_seqs = list(dated_clus_met["strain"])
@@ -514,8 +518,6 @@ for clus in clus_to_run:
     clus_data["observed_countries"] = observed_countries
     # Use below to print observed countries, if interested
     #print(f"The cluster is found in: {observed_countries}\n")
-    if clus != "S222":
-        print("Remember, countries are not set for clusters other than S222")
 
     # Let's get some summary stats on number of sequences, first, and last, for each country.
     country_info, country_dates = get_summary(cluster_meta, observed_countries)
@@ -552,7 +554,8 @@ for clus in clus_to_run:
             alert_first_date[clus] = before_date
 
     # Make a version of Delta which does not have as much UK/India sequences for increased viewability
-    if clus == "21AS478":
+    #if clus == "21AS478":
+    if clus_build_name == "21A.Delta":
         nouk_delta_meta = cluster_meta[
             cluster_meta["country"].apply(lambda x: x != "United Kingdom" and x != "India")
         ]
@@ -586,7 +589,8 @@ for clus in clus_to_run:
             copyfile(noUK_clusterlist_output, copypath2)
 
     # Make a version of Delta for Europe
-    if clus == "21AS478":
+    #if clus == "21AS478":
+    if clus_build_name == "21A.Delta":
         eu_delta_meta = cluster_meta[
             cluster_meta["region"].apply(lambda x: x == "Europe")
         ]
@@ -620,7 +624,8 @@ for clus in clus_to_run:
             copyfile(eu_clusterlist_output, copypath2)
 
     # Make a version of  Alpha-Delta which only have Swiss sequences for increased focus
-    if clus in ["501YV1", "501YV2", "501YV3", "21AS478"]:
+    #if clus in ["501YV1", "501YV2", "501YV3", "21AS478"]:
+    if clus_build_name in ["20I.Alpha.V1", "20H.Beta.V2", "20J.Gamma.V3", "21A.Delta"]:
         swiss_voc_meta = cluster_meta[
             cluster_meta["country"].apply(lambda x: x == "Switzerland")
         ]
@@ -656,7 +661,7 @@ for clus in clus_to_run:
 
     #######
     # print out the table
-    table_file = f"{tables_path}{clus_display}_table.tsv"
+    table_file = f"{tables_path}{clus_build_name}_table.tsv"
     ordered_country = country_info_df.sort_values(by="first_seq")
     clus_data["country_info_ordered"] = ordered_country
 
@@ -669,59 +674,8 @@ for clus in clus_to_run:
             ordered_country.to_csv(overall_tables_file, sep="\t", mode="a")
         mrk_tbl = ordered_country.to_markdown()
 
-        # url_params = "f_region=Europe"
-        # if 'url_params' in clusters[clus]:
-        #    url_params = clusters[clus]['url_params']
         nextstrain_url = clusters[clus]["nextstrain_url"]
-
-        ##################################
-        ##################################
-        #### Print the data tables in Markdown
-
-        # don't print DanishCluster in 'all tables'
-        # only print 'all tables' if running 'all clusters'
-        if "all" in clus_answer and clus != "DanishCluster":
-            with open(f"{tables_path}all_tables.md", "a") as fh:
-                fh.write(f"\n\n## {display_cluster}\n")
-                fh.write(f"[Focal Build]({nextstrain_url})\n\n")
-                if clus == "S477":
-                    fh.write(
-                        f"Note any pre-2020 Cambodian sequences are from SARS-like viruses in bats (not SARS-CoV-2).\n"
-                    )
-                if clus == "S501":
-                    fh.write(
-                        f"Note any pre-2020 Chinese sequences are from SARS-like viruses in bats (not SARS-CoV-2).\n"
-                    )
-                    fh.write(
-                        f"Note that this mutation has multiple amino-acid mutants - these numbers "
-                        "refer to _all_ these mutations (Y, S, T).\n"
-                    )
-                fh.write(mrk_tbl)
-                fh.write("\n\n")
-                fh.write(
-                    f"![Overall trends {clus_display}](/overall_trends_figures/overall_trends_{clus_display}.png)"
-                )
-
-        with open(f"{tables_path}{clus_display}_table.md", "w") as fh:
-            fh.write(f"\n\n## {display_cluster}\n")
-            fh.write(f"[Focal Build]({nextstrain_url})\n\n")
-            if clus == "S477":
-                fh.write(
-                    f"Note any pre-2020 Cambodian sequences are from SARS-like viruses in bats (not SARS-CoV-2).\n"
-                )
-            if clus == "S501":
-                fh.write(
-                    f"Note any pre-2020 Chinese sequences are from SARS-like viruses in bats (not SARS-CoV-2).\n"
-                )
-                fh.write(
-                    f"Note that this mutation has multiple amino-acid mutants - these numbers "
-                    "refer to _all_ these mutations (Y, S, T).\n"
-                )
-            fh.write(mrk_tbl)
-            fh.write("\n\n")
-            fh.write(
-                f"![Overall trends {clus_display}](/overall_trends_figures/overall_trends_{clus_display}.png)"
-            )
+        n_done += 1
 
     if print_acks:
         # remove all but EPI_ISL, on request from GISAID
@@ -729,17 +683,17 @@ for clus in clus_to_run:
         acknowledgement_table = cluster_meta.loc[:, ["gisaid_epi_isl"]]
         # do not put in acknowledgement folder, on request from GISAID
         # acknowledgement_table.to_csv(f'{acknowledgement_folder}{clus}_acknowledgement_table.tsv', sep="\t")
-        if clus != "DanishCluster":
+        if clus_build_name != "DanishCluster":
             acknowledgement_by_variant["acknowledgements"][
-                clus_display
+                clus_build_name
             ] = cluster_meta.loc[:, ["gisaid_epi_isl"]]["gisaid_epi_isl"].tolist()
 
         # only do this for 'all' runs as otherwise the main file won't be updated.
-        if clus != "DanishCluster" and "all" in clus_answer:
-            ack_out_folder = acknowledgement_folder_new + f"{clus_display}/"
+        if clus_build_name != "DanishCluster" and "all" in clus_answer:
+            ack_out_folder = acknowledgement_folder_new + f"{clus_build_name}/"
             if not os.path.exists(ack_out_folder):
                 os.mkdir(ack_out_folder)
-            ack_list = acknowledgement_by_variant["acknowledgements"][clus_display]
+            ack_list = acknowledgement_by_variant["acknowledgements"][clus_build_name]
             chunk_size = 1000
             chunks = [
                 ack_list[i : i + chunk_size]
@@ -748,8 +702,8 @@ for clus in clus_to_run:
 
             # get number & file names
             ack_file_names = ["{0:03}".format(i) for i in range(len(chunks))]
-            acknowledgement_keys["acknowledgements"][clus_display] = {}
-            acknowledgement_keys["acknowledgements"][clus_display]["numChunks"] = len(
+            acknowledgement_keys["acknowledgements"][clus_build_name] = {}
+            acknowledgement_keys["acknowledgements"][clus_build_name]["numChunks"] = len(
                 chunks
             )
 
@@ -844,19 +798,19 @@ if division:
             counts_by_2week = week2_counts.to_dict()
             division_all_sequence_counts[sel_coun][div]['2week'] = counts_by_2week
 
-
+t1 = time.time()
+print(f"Gathering up sequences took {round((t1-t0)/60,1)} min to run (partway through binning)\n")
 
 ######################################################################################################
 ##################################
 #### PREPARING FOR OF PLOTTING - putting clusters into date bins
-
 
 for clus in clus_to_run:
     print(f"\nPutting {clus} data into bins by date\n")
 
     clus_data = clusters[clus]
     wanted_seqs = clus_data["wanted_seqs"]
-    clus_display = clus_data["build_name"]
+    clus_build_name = clus_data["build_name"]
     cluster_meta = clus_data["cluster_meta"]
     observed_countries = clus_data["observed_countries"]
     country_dates = clus_data["country_dates"]
@@ -992,7 +946,7 @@ for clus in clus_to_run:
     c_i[c_i["num_seqs"] > cutoff_num_seqs]
     #print(f"Countries with >{cutoff_num_seqs} seqs in cluster {clus}:")
     #print("\t", ", ".join(c_i[c_i["num_seqs"] > 10].index))
-    if len(c_i[c_i["num_seqs"] > 10]) > 0 and clus != "DanishCluster":
+    if len(c_i[c_i["num_seqs"] > 10]) > 0 and clus_build_name != "DanishCluster":
         clusters_tww.append(clus)
     print("")
 
@@ -1038,13 +992,14 @@ t0 = time.time()
 
 countries_plotted = {}
 
+ndone = 1
 for clus in clus_to_run:
 
-    print(f"\nPlotting & writing out cluster {clus}\n")
+    print(f"\nPlotting & writing out cluster {clus}: number {ndone} of {len(clus_to_run)}")
 
     clus_data = clusters[clus]
     wanted_seqs = clus_data["wanted_seqs"]
-    clus_display = clus_data["build_name"]
+    clus_build_name = clus_data["build_name"]
     cluster_meta = clus_data["cluster_meta"]
     observed_countries = clus_data["observed_countries"]
     country_dates = clus_data["country_dates"]
@@ -1057,9 +1012,7 @@ for clus in clus_to_run:
     smoothing /= smoothing.sum()
 
     # Only plot countries with >= X seqs
-    min_to_plot = 400
-    # if clus == "S222":
-    #    min_to_plot = 200
+    min_to_plot = cutoff_num_seqs
 
     countries_to_plot_min = country_info_df[
         country_info_df.num_seqs > min_to_plot
@@ -1074,220 +1027,51 @@ for clus in clus_to_run:
         if x in countries_to_plot_final
     ]
 
-    if len(countries_to_plot) > len(country_styles_all):
-        print("\nWARNING!! NOT ENOUGH COLORS FOR PLOTTING!")
-
-    if clus == "S222a":
-        country_styles_custom = country_styles
-    else:
-        unused_countries = [
-            x for x in country_styles_all if x not in countries_to_plot
-        ]
-        country_styles_custom = {}
-        for x in countries_to_plot:
-            if x in country_styles_all.keys():
-                country_styles_custom[x] = country_styles_all[x]
-            else:
-                country_styles_custom[x] = country_styles_all[unused_countries.pop(0)]
-
-    # Make a plot
-    repeat = 1
-    if clus == "S222":
-        repeat = 2
-
-    while repeat > 0:
-
-        # fig = plt.figure(figsize=(10,5))
-        # fig, axs=plt.subplots(1,1, figsize=(10,5))
-        fs = 14
-        # fig, (ax1, ax2, ax3) = plt.subplots(nrows=3, sharex=True,figsize=(10,7),
-        #                                    gridspec_kw={'height_ratios':[1,1,3]})
-        # Change to just show Travel to spain only. see above for old 3 panel version
-        if repeat == 2:
-            fig, (ax1, ax3) = plt.subplots(
-                nrows=2,
-                sharex=True,
-                figsize=(10, 6),
-                gridspec_kw={"height_ratios": [1, 3]},
-            )
-        else:
-            fig, ax3 = plt.subplots(1, 1, figsize=(10, 5), dpi=72)
-
-        if repeat == 2:
-            i = 0
-            for coun in travel_order:
-                if coun in q_free_to_spain:
-                    q_times = q_free_to_spain[coun]
-                    strt = datetime.datetime.strptime(q_times["start"], "%Y-%m-%d")
-                    end = datetime.datetime.strptime(q_times["end"], "%Y-%m-%d")
-                    y_start = i * 0.022
-                    height = 0.02
-                    ax1.add_patch(
-                        Rectangle(
-                            (strt, y_start),
-                            end - strt,
-                            height,
-                            ec=country_styles_custom[coun]["c"],
-                            fc=country_styles_custom[coun]["c"],
-                        )
-                    )
-
-                    ax1.text(strt, y_start + 0.003, q_times["msg"], fontsize=fs * 0.8)
-                    if coun == "Denmark":
-                        strt = datetime.datetime.strptime(
-                            q_free_to_spain["Denmark2"]["start"], "%Y-%m-%d"
-                        )
-                        end = datetime.datetime.strptime(
-                            q_free_to_spain["Denmark2"]["end"], "%Y-%m-%d"
-                        )
-                        ax1.add_patch(
-                            Rectangle(
-                                (strt, y_start),
-                                end - strt,
-                                height,
-                                ec=country_styles_custom[coun]["c"],
-                                fc="none",
-                                hatch="/",
-                            )
-                        )
-                        ax1.text(
-                            strt,
-                            y_start + 0.003,
-                            q_free_to_spain["Denmark2"]["msg"],
-                            fontsize=fs * 0.8,
-                        )
-                i = i + 1
-            ax1.set_ylim([0, y_start + height])
-            ax1.text(
-                datetime.datetime.strptime("2020-05-03", "%Y-%m-%d"),
-                y_start,
-                "Quarantine-free",
-                fontsize=fs,
-            )
-            ax1.text(
-                datetime.datetime.strptime("2020-05-03", "%Y-%m-%d"),
-                y_start - height - 0.005,
-                "Travel to/from Spain",
-                fontsize=fs,
-            )
-            ax1.text(
-                datetime.datetime.strptime("2020-05-03", "%Y-%m-%d"),
-                y_start - height - height - 0.01,
-                "(on return)",
-                fontsize=fs,
-            )
-            ax1.get_yaxis().set_visible(False)
-
-        # for a simpler plot of most interesting countries use this:
-        for coun in [x for x in countries_to_plot_min]:
-            (
-                week_as_date,
-                cluster_count,
-                total_count,
-                unsmoothed_cluster_count,
-                unsmoothed_total_count,
-            ) = non_zero_counts(cluster_data, total_data, coun, smoothing=smoothing)
-            # remove last data point if that point as less than frac sequences compared to the previous count
-            week_as_date, cluster_count, total_count = trim_last_data_point(
-                week_as_date, cluster_count, total_count, frac=0.1, keep_count=10
-            )
-            if len(cluster_count) < len(
-                unsmoothed_cluster_count
-            ):  # if the trim_last_data_point came true, match trimming
-                unsmoothed_cluster_count = unsmoothed_cluster_count[:-1]
-                unsmoothed_total_count = unsmoothed_total_count[:-1]
-
-            json_output[clus_display][coun] = {}
-            json_output[clus_display][coun]["week"] = [
-                datetime.datetime.strftime(x, "%Y-%m-%d") for x in week_as_date
-            ]
-            json_output[clus_display][coun]["total_sequences"] = [
-                int(x) for x in total_count
-            ]
-            json_output[clus_display][coun]["cluster_sequences"] = [
-                int(x) for x in cluster_count
-            ]
-            json_output[clus_display][coun]["unsmoothed_cluster_sequences"] = [
-                int(x) for x in unsmoothed_cluster_count
-            ]
-            json_output[clus_display][coun]["unsmoothed_total_sequences"] = [
-                int(x) for x in unsmoothed_total_count
-            ]
-
-            # This used to only plot a subset (those in 'countries to plot' below.)
-            # However for a long time (as of Jul 21) all have been selected,
-            # And I think this is most intuitive - so setting all to 'True'
-            countries_plotted[coun] = "True" #"False"
-
-            if coun in countries_to_plot:
-                ax3.plot(
-                    week_as_date,
-                    cluster_count / total_count,
-                    color=country_styles_custom[coun]["c"],
-                    linestyle=country_styles_custom[coun]["ls"],
-                    label=coun,
-                )
-                ax3.scatter(
-                    week_as_date,
-                    cluster_count / total_count,
-                    s=[marker_size(n) for n in unsmoothed_total_count],
-                    color=country_styles_custom[coun]["c"],
-                    linestyle=country_styles_custom[coun]["ls"],
-                )  #
-                countries_plotted[coun] = "True"
-
-        for ni, n in enumerate([0, 1, 3, 10, 30, 100]):
-            ax3.scatter(
-                [week_as_date[0]],
-                [0.08 + ni * 0.07],
-                s=marker_size(n + 0.1),
-                edgecolor="k",
-                facecolor="w",
-            )
-            ax3.text(week_as_date[1], 0.06 + ni * 0.07, f"n>{n}" if n else "n=1")
-            #          color=country_styles[coun]['c'], linestyle=country_styles[coun]['ls'], label=coun)
-
-        ax3.text(datetime.datetime(2020, 10, 1), 0.9, f"{clus_display}", fontsize=fs)
-        plt.legend(ncol=1, fontsize=fs * 0.8, loc=2)
-        fig.autofmt_xdate(rotation=30)
-        ax3.tick_params(labelsize=fs * 0.8)
-        ax3.set_ylabel("frequency", fontsize=fs)
-        max_date = country_info_df["last_seq"].max()
-        ax3.set_ylim(0, 1)
-        ax3.set_xlim(
-            datetime.datetime(2020, 5, 1),
-            datetime.datetime.strptime(max_date, "%Y-%m-%d"),
+    for coun in [x for x in countries_to_plot_min]:
+        (
+            week_as_date,
+            cluster_count,
+            total_count,
+            unsmoothed_cluster_count,
+            unsmoothed_total_count,
+        ) = non_zero_counts(cluster_data, total_data, coun, smoothing=smoothing)
+        # remove last data point if that point as less than frac sequences compared to the previous count
+        week_as_date, cluster_count, total_count = trim_last_data_point(
+            week_as_date, cluster_count, total_count, frac=0.1, keep_count=10
         )
-        plt.show()
-        plt.tight_layout()
+        if len(cluster_count) < len(
+            unsmoothed_cluster_count
+        ):  # if the trim_last_data_point came true, match trimming
+            unsmoothed_cluster_count = unsmoothed_cluster_count[:-1]
+            unsmoothed_total_count = unsmoothed_total_count[:-1]
 
-        # spain opens borders
-        if clus == "S222":
-            ax3.text(
-                datetime.datetime.strptime("2020-06-21", "%Y-%m-%d"),
-                0.05,
-                "Spain opens borders",
-                rotation="vertical",
-                fontsize=fs * 0.8,
-            )
+        json_output[clus_build_name][coun] = {}
+        json_output[clus_build_name][coun]["week"] = [
+            datetime.datetime.strftime(x, "%Y-%m-%d") for x in week_as_date
+        ]
+        json_output[clus_build_name][coun]["total_sequences"] = [
+            int(x) for x in total_count
+        ]
+        json_output[clus_build_name][coun]["cluster_sequences"] = [
+            int(x) for x in cluster_count
+        ]
+        json_output[clus_build_name][coun]["unsmoothed_cluster_sequences"] = [
+            int(x) for x in unsmoothed_cluster_count
+        ]
+        json_output[clus_build_name][coun]["unsmoothed_total_sequences"] = [
+            int(x) for x in unsmoothed_total_count
+        ]
 
-        travel = ""
-        if repeat == 2:
-            travel = "Travel"
+        # This used to only plot a subset (those in 'countries to plot' below.)
+        # However for a long time (as of Jul 21) all have been selected,
+        # And I think this is most intuitive - so setting all to 'True'
+        countries_plotted[coun] = "True" #"False"
+
+        ndone += 1
+
         if print_files:
-            plt.savefig(figure_path + f"overall_trends_{clus_display}{travel}.{fmt}")
-            trends_path = figure_path + f"overall_trends_{clus_display}{travel}.{fmt}"
-            copypath = trends_path.replace(
-                f"trends_{clus_display}{travel}",
-                "trends-{}".format(datetime.date.today().strftime("%Y-%m-%d")),
-            )
-            copyfile(trends_path, copypath)
-
-        repeat = repeat - 1
-
-        if print_files:
-            with open(tables_path + f"{clus_display}_data.json", "w") as fh:
-                json.dump(json_output[clus_display], fh)
+            with open(tables_path + f"{clus_build_name}_data.json", "w") as fh:
+                json.dump(json_output[clus_build_name], fh)
 
 if "all" in clus_answer:
     for coun in countries_plotted.keys():
@@ -1326,27 +1110,7 @@ print("############################\n\n")
 # Start of the country data/plotting
 
 
-# ask user if they want to continue (if something went wrong above, might not want to)
-do_country = False
-if "all" in clus_answer:
-    print_answer = input("\nContinue to country plotting? (y/n) (Enter is no): ")
-    if print_answer in ["y", "Y", "yes", "YES", "Yes"]:
-        do_country = True
-else:
-    print("Can't do country plot as didn't do 'all' clusters")
 
-if do_country == False:
-    print(
-        "You can alway run this step by calling `plot_country_data(clusters, proposed_coun_to_plot, print_files)`"
-    )
-
-do_divisions_country = False
-if "all" in clus_answer:
-    print_answer = input(
-        "\nContinue to USA- & Swiss-specific country plotting? (y/n) (Enter is no): "
-    )
-    if print_answer in ["y", "Y", "yes", "YES", "Yes"]:
-        do_divisions_country = True
         
 
 def get_ordered_clusters_to_plot(clusters, division=False, selected_country=None):
@@ -1425,24 +1189,17 @@ def plot_country_data(
 ):
     country_week = {clus: {} for clus in clusters}
 
-    fs = 14
-    rws = int(np.ceil((len(proposed_coun_to_plot) + 1) / 2))
-    fig, axs = plt.subplots(
-        nrows=rws, ncols=2, sharex=True, figsize=(9, 11)  # len(countries_to_plot)+1,
-    )
 
     min_week = datetime.datetime(2020, 12, 31)
     max_week = datetime.datetime(2020, 1, 1)
     week_as_dates = {}
     json_output = {}
     json_output["countries"] = {}
-    ptchs = {}
 
-    for coun, ax in zip(proposed_coun_to_plot, fig.axes[1:]):  # axs[1:]):
+    for coun in proposed_coun_to_plot:
         i = 0
         first_clus_count = []
 
-        # json_output["countries"][coun] = {"week": {}, "total_sequences": {}}
         country_data = {"week": {}, "total_sequences": {}}
 
         for clus in clus_keys:
@@ -1457,15 +1214,6 @@ def plot_country_data(
                 total_data = clusters[clus]["total_data_2wk"]
 
             if coun not in cluster_data:
-                if clus == clus_keys[-1]:
-                    ax.fill_between(
-                        week_as_date,
-                        cluster_count / total_count,
-                        1,
-                        facecolor=grey_color,
-                    )
-                    patch = mpatches.Patch(color=grey_color, label=f"other")
-                    ptchs["other"] = patch
                 i += 1
                 continue
             (
@@ -1507,23 +1255,6 @@ def plot_country_data(
                 first_clus_count = [0] * len(total_count)
             cluster_count = first_clus_count + cluster_count  # unindented
 
-            ax.fill_between(
-                week_as_date,
-                first_clus_count / total_count,
-                cluster_count / total_count,
-                facecolor=clusters[clus]["col"],
-            )
-            patch = mpatches.Patch(color=clusters[clus]["col"], label=lab)
-
-            if clus not in ptchs:
-                ptchs[clus] = patch
-
-            if clus == clus_keys[-1]:
-                ax.fill_between(
-                    week_as_date, cluster_count / total_count, 1, facecolor=grey_color
-                )
-                patch = mpatches.Patch(color=grey_color, label=f"other")
-
             first_clus_count = cluster_count  # unindented
             i += 1
         country_data["week"] = [
@@ -1535,9 +1266,6 @@ def plot_country_data(
         if len(total_count)>=2:
             json_output["countries"][coun] = country_data
 
-        ax.text(datetime.datetime(2020, 6, 1), 0.7, coun, fontsize=fs)
-        ax.tick_params(labelsize=fs * 0.8)
-
     json_output["plotting_dates"] = {}
     json_output["plotting_dates"]["min_date"] = datetime.datetime.strftime(
         min_week, "%Y-%m-%d"
@@ -1546,25 +1274,9 @@ def plot_country_data(
         max_week, "%Y-%m-%d"
     )
 
-    fig.axes[0].legend(handles=ptchs.values(), loc=3, fontsize=fs * 0.7, ncol=3)
-    fig.axes[0].axis("off")
-    fig.autofmt_xdate(rotation=30)
-    plt.show()
-    plt.tight_layout()
-
     if print_files:
-        fmt = "pdf"
         with open(tables_path + f"{file_prefix}_data.json", "w") as fh:
             json.dump(json_output, fh)
-
-        plt.savefig(figure_path + f"{file_prefix}_compare.png")
-
-        plt.savefig(figure_only_path + f"{file_prefix}_compare.{fmt}")
-        trends_path = figure_only_path + f"{file_prefix}_compare.{fmt}"
-        copypath = trends_path.replace(
-            "compare", "compare-{}".format(datetime.date.today().strftime("%Y-%m-%d"))
-        )
-        copyfile(trends_path, copypath)
 
 
 if do_country: